+++ angular-workspace/components/slide-toggle/slide-toggle.stories.ts
@@ -1,0 +1,20 @@
+import { storiesOf } from "@storybook/angular";
+
+import { storiesOfSlideToggle } from "dso-toolkit";
+import { DsoSlideToggle } from "../../projects/component-library/src/public-api";
+import { templateContainer } from "../../templates";
+
+import readme from "./readme.md";
+
+storiesOfSlideToggle({
+  parameters: {
+    module,
+    storiesOf,
+    readme,
+    storyApiOptions: {
+      parameters: [{ component: DsoSlideToggle }],
+    },
+  },
+  templateContainer,
+  storyTemplates: ({ slideToggleTemplate }) => ({ slideToggleTemplate }),
+});