+++ packages/core/src/components/slide-toggle/slide-toggle.interfaces.ts
@@ -1,0 +1,5 @@
+export interface SlideToggleActiveEvent {
+  originalEvent?: Event;
+  /** The value that was set */
+  checked: boolean;
+}