+++ packages/dso-toolkit/src/components/slide-toggle/slide-toggle.models.ts
@@ -1,0 +1,14 @@
+export interface SlideToggle {
+  dsoActiveChange: (e: CustomEvent<SlideToggleChangeEvent>) => void;
+  checked: boolean;
+  disabled?: boolean;
+  accessibleLabel?: string;
+  labelledbyId?: string;
+  label?: string;
+  useOwnLabelId?: string;
+}
+
+export interface SlideToggleChangeEvent {
+  originalEvent?: Event;
+  checked: boolean;
+}